requirements.txt
# API framework
fastapi>=0.115.0
uvicorn[standard]>=0.30.0

# LLM providers
anthropic>=0.84.0
google-genai>=1.0.0

# Embeddings and vector store
sentence-transformers>=3.0.0
faiss-cpu>=1.8.0

# Document processing
pdfplumber>=0.11.0
beautifulsoup4>=4.12.0
httpx>=0.27.0

# Web search
duckduckgo-search>=6.0.0

# Frontend
streamlit>=1.40.0

# Utilities
python-dotenv>=1.0.0
pydantic-settings>=2.5.0
python-multipart>=0.0.9
